feat(data-resource): change blake3 hash to be of type Vec<u64>

+++ data-resource/src/blake3.rs
@@ -31,7 +31,7 @@ pub struct ResourceId {}
 #[derive(
     Hash, Ord, PartialOrd, Eq, PartialEq, Clone, Debug, Serialize, Deserialize,
 )]
-pub struct Hash(Vec<u8>);
+pub struct Hash(Vec<u64>);
 
 impl Display for Hash {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
